requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-appconf==1.1.0
django-compressor==4.5.1
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
tzdata==2025.2
